+++ src/test/java/lotto/model/LottoTest.java
@@ -1,4 +1,4 @@
-package lotto.controller;
+package lotto.model;
 
 import static org.assertj.core.api.Assertions.*;
