+++ src/test/java/lotto/model/LottosTest.java
@@ -1,4 +1,4 @@
-package lotto.controller;
+package lotto.model;
 
 import static org.assertj.core.api.Assertions.*;
